webpack модуляция
