+++ src/components/show-more-button.js
@@ -1,0 +1,5 @@
+const createShowMoreButtonTemplate = () => {
+  return `<button class="films-list__show-more">Show more</button>`;
+};
+
+export {createShowMoreButtonTemplate};
